requirements.txt
numpy>=1.21
pandas>=1.3
biopython>=1.79
torch>=1.10
scipy>=1.7
# openmm>=7.6
# pdbfixer
transformers
openai
python-dotenv
jupyter
matplotlib
seaborn
